Fix more compile errors.

+++ server/src/main/java/org/elasticsearch/common/lucene/search/function/CombineFunction.java
@@ -38,10 +38,10 @@ public enum CombineFunction implements Writeable {
         public Explanation explain(Explanation queryExpl, Explanation funcExpl, float maxBoost) {
             Explanation boostExpl = Explanation.match(maxBoost, "maxBoost");
             Explanation minExpl = Explanation.match(
-                    Math.min(funcExpl.getValue(), maxBoost),
+                    Math.min(funcExpl.getValue().floatValue(), maxBoost),
                     "min of:",
                     funcExpl, boostExpl);
-            return Explanation.match(queryExpl.getValue() * minExpl.getValue(),
+            return Explanation.match(queryExpl.getValue().floatValue() * minExpl.getValue().floatValue(),
                     "function score, product of:", queryExpl, minExpl);
         }
     },
@@ -55,7 +55,7 @@ public enum CombineFunction implements Writeable {
         public Explanation explain(Explanation queryExpl, Explanation funcExpl, float maxBoost) {
             Explanation boostExpl = Explanation.match(maxBoost, "maxBoost");
             return Explanation.match(
-                    Math.min(funcExpl.getValue(), maxBoost),
+                    Math.min(funcExpl.getValue().floatValue(), maxBoost),
                     "min of:",
                     funcExpl, boostExpl);
         }
@@ -69,9 +69,9 @@ public enum CombineFunction implements Writeable {
 
         @Override
         public Explanation explain(Explanation queryExpl, Explanation funcExpl, float maxBoost) {
-            Explanation minExpl = Explanation.match(Math.min(funcExpl.getValue(), maxBoost), "min of:",
+            Explanation minExpl = Explanation.match(Math.min(funcExpl.getValue().floatValue(), maxBoost), "min of:",
                     funcExpl, Explanation.match(maxBoost, "maxBoost"));
-            return Explanation.match(Math.min(funcExpl.getValue(), maxBoost) + queryExpl.getValue(), "sum of",
+            return Explanation.match(Math.min(funcExpl.getValue().floatValue(), maxBoost) + queryExpl.getValue().floatValue(), "sum of",
                     queryExpl, minExpl);
         }
 
@@ -84,10 +84,10 @@ public enum CombineFunction implements Writeable {
 
         @Override
         public Explanation explain(Explanation queryExpl, Explanation funcExpl, float maxBoost) {
-            Explanation minExpl = Explanation.match(Math.min(funcExpl.getValue(), maxBoost), "min of:",
+            Explanation minExpl = Explanation.match(Math.min(funcExpl.getValue().floatValue(), maxBoost), "min of:",
                     funcExpl, Explanation.match(maxBoost, "maxBoost"));
             return Explanation.match(
-                    (float) ((Math.min(funcExpl.getValue(), maxBoost) + queryExpl.getValue()) / 2.0), "avg of",
+                    (float) ((Math.min(funcExpl.getValue().floatValue(), maxBoost) + queryExpl.getValue().floatValue()) / 2.0), "avg of",
                     queryExpl, minExpl);
         }
 
@@ -101,10 +101,10 @@ public enum CombineFunction implements Writeable {
         @Override
         public Explanation explain(Explanation queryExpl, Explanation funcExpl, float maxBoost) {
             Explanation innerMinExpl = Explanation.match(
-                    Math.min(funcExpl.getValue(), maxBoost), "min of:",
+                    Math.min(funcExpl.getValue().floatValue(), maxBoost), "min of:",
                     funcExpl, Explanation.match(maxBoost, "maxBoost"));
             return Explanation.match(
-                    Math.min(Math.min(funcExpl.getValue(), maxBoost), queryExpl.getValue()), "min of",
+                    Math.min(Math.min(funcExpl.getValue().floatValue(), maxBoost), queryExpl.getValue().floatValue()), "min of",
                     queryExpl, innerMinExpl);
         }
 
@@ -118,10 +118,10 @@ public enum CombineFunction implements Writeable {
         @Override
         public Explanation explain(Explanation queryExpl, Explanation funcExpl, float maxBoost) {
             Explanation innerMinExpl = Explanation.match(
-                    Math.min(funcExpl.getValue(), maxBoost), "min of:",
+                    Math.min(funcExpl.getValue().floatValue(), maxBoost), "min of:",
                     funcExpl, Explanation.match(maxBoost, "maxBoost"));
             return Explanation.match(
-                    Math.max(Math.min(funcExpl.getValue(), maxBoost), queryExpl.getValue()), "max of:",
+                    Math.max(Math.min(funcExpl.getValue().floatValue(), maxBoost), queryExpl.getValue().floatValue()), "max of:",
                     queryExpl, innerMinExpl);
         }
